Slight reorganization of fastxpub files
