+++ fastxpub/additional_sources/rand-mock.c
